% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{charFormatOS}
\alias{charFormatOS}
\title{A Function for formatting fields as character}
\usage{
charFormatOS(input)
}
\arguments{
\item{input}{input (character, numeric or integer)}
}
\value{
sprintfed input
}
\description{
This function sprintfs the fields as character "%-8s". Mainly a helper function for dfTolines
}
\examples{



}
