% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dfTolines}
\alias{dfTolines}
\title{A Function for printing a data frame into lines}
\usage{
dfTolines(df, funct = charFormatOS)
}
\arguments{
\item{df}{a data frame}

\item{funct}{a format function. charFormatOs is the default, and recommended option for now.}
}
\value{
lines
}
\description{
This function allows you transform fem expanded lines into a data frame
}
\examples{



}
