% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expandedlines2bulk}
\alias{expandedlines2bulk}
\title{A Function for writing bulk data strings from expanded lines}
\usage{
expandedlines2bulk(lines)
}
\arguments{
\item{lines}{expanded lines.}
}
\value{
a string in bulk data format.
}
\description{
This function allows you to use expanded lines and rerturn bulk data strings. Continuation cards must start with "+" for this function to work.
}
\examples{


}
