% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{femtoCaseBulk}
\alias{femtoCaseBulk}
\title{A Function for reading fem files}
\usage{
femtoCaseBulk(file)
}
\arguments{
\item{file}{file to read}
}
\value{
a character vector of case, bulk, and comments after enddata. Size <=3
}
\description{
This function allows you to read fem files, returning a character vector of size 3: case, bulk and comments after ENDDATA.
}
\examples{


}
