% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCardName}
\alias{getCardName}
\title{A Function for extracting card images from lines}
\usage{
getCardName(lines)
}
\arguments{
\item{lines}{character vector with expanded bulk lines}
}
\value{
a character vector card images/names
}
\description{
This function allows you remove commented lines, returning a character vector of the non-commented lines
}
\examples{


}
