% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{linesToDF}
\alias{linesToDF}
\title{A Function for reading lines into a data frame}
\usage{
linesToDF(lines, numberOfFields = 9)
}
\arguments{
\item{lines}{expanded lines}

\item{numberOfFields}{number of fields}
}
\value{
a data frame
}
\description{
This function allows you transform fem expanded lines into a data frame
}
\examples{


}
