% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFemLines.R
\name{readFemLines}
\alias{readFemLines}
\title{A Function for reading Optistruct Files}
\usage{
readFemLines(file, nchars)
}
\arguments{
\item{file}{file to read}

\item{nchars}{number of characters}
}
\value{
a character vector with Optistruct cards
}
\description{
This function allows you to read the first characters (nchars) of a file, returning a character vector of optistruct cards
}
\examples{


}
