% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{removeComments}
\alias{removeComments}
\title{A Function for removing commented lines from a character vector of expanded bulk lines.}
\usage{
removeComments(lines)
}
\arguments{
\item{lines}{character vector with the bulk lines}
}
\value{
a character vector of non-commented bulk lines.
}
\description{
This function allows you remove commented lines, returning a character vector of the non-commented lines
}
\examples{


}
