% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shortFormatOS}
\alias{shortFormatOS}
\title{A Function for formatting fields depending on its type}
\usage{
shortFormatOS(input)
}
\arguments{
\item{input}{input (character, numeric or integer)}
}
\value{
sprintfed input
}
\description{
This function sprintfs the fields depending on its type. Mainly a helper function for dfTolines
}
\examples{



}
