% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{writeInclude}
\alias{writeInclude}
\title{A Function for writing a master and include files from a merged fem and a vector of card names}
\usage{
writeInclude(file, ocards, includename, mastername)
}
\arguments{
\item{ocards}{cards to be moved}

\item{includename}{includename}

\item{mastername}{mastername}
}
\value{
Nothing
}
\description{
This function allows you split a merged fem into a master plus a include file with the selected type of cards
}
\examples{


}
