% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{linesToDF}
\alias{linesToDF}
\title{A Function for}
\usage{
linesToDF(lines, numberOfFields)
}
\arguments{
\item{file}{fem file}

\item{mastername}{name of the master to write}

\item{sorted}{if the cards need to be sorted in the include files (to diff, mostly)}
}
\value{
Nothing
}
\description{
This function allows you
}
\examples{


}
