% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFemLines.R, R/writeFemLines.R
\name{readFemLines}
\alias{readFemLines}
\alias{readFemLines}
\title{A Function for reading Optistruct Files}
\usage{
readFemLines(file, nchars)

readFemLines(file, nchars)
}
\arguments{
\item{file}{file to read}

\item{nchars}{number of characters}

\item{file}{file to write}

\item{x}{character vector}
}
\value{
a character vector with Optistruct cards

nothing
}
\description{
This function allows you to read the first characters (nchars) of a file, returning a character vector of optistruct cards

This function allows you to write an OptiStruct file from a character vector of the kind returned from readFemLines.R
}
\examples{




}
